Change `computeAddress` Visibility to `internal`
This PR changes the visibility of the `computeAddress` library functions to be `internal`. Library functions that are `public` generate code to call to an external library contract.

This causes some inconveniences, for example uses of `computeAddress` with `forge script` to potentially broadcast a transaction to deploy the `SafeSingletonDeployer` library (even if it isn't actually needed).